float textsize

+++ Adafruit_GFX.h
@@ -112,17 +112,17 @@ public:
   void drawRGBBitmap(int16_t x, int16_t y, uint16_t *bitmap, uint8_t *mask,
                      int16_t w, int16_t h);
   void drawChar(int16_t x, int16_t y, unsigned char c, uint16_t color,
-                uint16_t bg, uint8_t size);
+                uint16_t bg, float size);
   void drawChar(int16_t x, int16_t y, unsigned char c, uint16_t color,
-                uint16_t bg, uint8_t size_x, uint8_t size_y);
+                uint16_t bg, float size_x, float size_y);
   void getTextBounds(const char *string, int16_t x, int16_t y, int16_t *x1,
                      int16_t *y1, uint16_t *w, uint16_t *h);
   void getTextBounds(const __FlashStringHelper *s, int16_t x, int16_t y,
                      int16_t *x1, int16_t *y1, uint16_t *w, uint16_t *h);
   void getTextBounds(const String &str, int16_t x, int16_t y, int16_t *x1,
                      int16_t *y1, uint16_t *w, uint16_t *h);
-  void setTextSize(uint8_t s);
-  void setTextSize(uint8_t sx, uint8_t sy);
+  void setTextSize(float s);
+  void setTextSize(float sx, float sy);
   void setFont(const GFXfont *f = NULL);
 
   /**********************************************************************/
@@ -244,8 +244,8 @@ protected:
   int16_t cursor_y;     ///< y location to start print()ing text
   uint16_t textcolor;   ///< 16-bit background color for print()
   uint16_t textbgcolor; ///< 16-bit text color for print()
-  uint8_t textsize_x;   ///< Desired magnification in X-axis of text to print()
-  uint8_t textsize_y;   ///< Desired magnification in Y-axis of text to print()
+  float textsize_x;     ///< Desired magnification in X-axis of text to print()
+  float textsize_y;     ///< Desired magnification in Y-axis of text to print()
   uint8_t rotation;     ///< Display rotation (0 thru 3)
   bool wrap;            ///< If set, 'wrap' text at right edge of display
   bool _cp437;          ///< If set, use correct CP437 charset (default is off)
@@ -302,8 +302,8 @@ private:
   Adafruit_GFX *_gfx;
   int16_t _x1, _y1; // Coordinates of top-left corner
   uint16_t _w, _h;
-  uint8_t _textsize_x;
-  uint8_t _textsize_y;
+  float _textsize_x;
+  float _textsize_y;
   uint16_t _outlinecolor, _fillcolor, _textcolor;
   char _label[10];
